Make test suites run common
