+++ testing/endtoend/kurtosis_e2e_helper.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"os/exec"
+	"slices"
 	"strconv"
 	"strings"
 	"testing"
@@ -11,6 +12,7 @@ import (
 
 	"github.com/OffchainLabs/prysm/v7/api/client/beacon"
 	"github.com/OffchainLabs/prysm/v7/config/params"
+	"github.com/OffchainLabs/prysm/v7/testing/endtoend/kurtosis"
 	"github.com/OffchainLabs/prysm/v7/testing/require"
 	"github.com/bazelbuild/rules_go/go/tools/bazel"
 )
@@ -19,12 +21,95 @@ const (
 	// ETHEREUM_PACKAGE is the identifier of the ethereum-package Starlark package used in these tests.
 	ETHEREUM_PACKAGE = "github.com/ethpandaops/ethereum-package"
 
+	// LATE_SYNC_NODE_DELAY is how long after genesis the skip_start sync nodes are
+	// started, so the chain has advanced and finalized: the normal-sync node then
+	// has history to catch up over P2P, and the checkpoint-sync node has a
+	// finalized checkpoint to sync from.
+	LATE_SYNC_NODE_DELAY = 6 * time.Minute
+
+	// SYNC_NODE_SERVICE is the skip_start node for the P2P (genesis) sync test.
+	SYNC_NODE_SERVICE = "cl-3-prysm-geth"
+
+	// CHECKPOINT_SYNC_NODE_SERVICE is the skip_start node for the checkpoint sync test.
+	CHECKPOINT_SYNC_NODE_SERVICE = "cl-4-prysm-geth"
+
 	BEACON_CHAIN_IMAGE_TARGET = "cmd/beacon-chain/oci_image_tarball_e2e/tarball.tar"
 	VALIDATOR_IMAGE_TARGET    = "cmd/validator/oci_image_tarball_e2e/tarball.tar"
 
 	BEACON_CHAIN_IMAGE_NAME = "gcr.io/offchainlabs/prysm/beacon-chain:latest"
 	VALIDATOR_IMAGE_NAME    = "gcr.io/offchainlabs/prysm/validator:latest"
 )
+
+type KurtosisTestSuites struct {
+	enclaveName string
+	configPath  string
+	epochsToRun uint64
+	runSyncTest bool
+}
+
+func (k *KurtosisTestSuites) Run(t *testing.T) {
+	// Note: Subtests can be run in parallel as they use separate enclaves.
+	t.Parallel()
+
+	ctx := t.Context()
+
+	kw, err := kurtosis.NewKurtosisWrapper(t, ctx, k.enclaveName)
+	require.NoError(t, err, "Failed to create Kurtosis wrapper")
+
+	require.NoError(t, kw.CreateEnclave(), "Failed to create Kurtosis enclave")
+	t.Cleanup(func() {
+		if t.Failed() {
+			// Dump logs so that we can see what went wrong before the enclave is destroyed.
+			kw.DumpFailedAssertoorLogs()
+		}
+		if err := kw.DestroyEnclave(); err != nil {
+			t.Logf("Failed to cleanup enclave: %v", err)
+		}
+	})
+
+	require.NoError(t, kw.RunPackageWithNetworkConfig(
+		ETHEREUM_PACKAGE,
+		k.configPath,
+	), "Failed to run ethereum package")
+
+	restURLs, err := kw.NewBeaconRESTEndpoints()
+	require.NoError(t, err, "Failed to resolve beacon REST endpoints")
+
+	// Create a beacon API client to
+	// 1. Fetch genesis information.
+	// 2. Fetch config spec for hydrating params.
+	client, err := beacon.NewClient(restURLs[0])
+	require.NoError(t, err, "Failed to create beacon API client")
+
+	// Gate on node readiness once, then every API call below is a single request.
+	waitForNodeReady(t, ctx, client)
+
+	// Hydrate params with the config the enclave is actually running, so
+	// the timeout below is computed against the real network config.
+	cfg := fetchConfig(t, ctx, client)
+	params.SetActiveTestCleanup(t, cfg)
+
+	// Set deadline for assertoor.
+	genesisTime := fetchGenesisTime(t, ctx, client)
+	secondsPerEpoch := uint64(params.BeaconConfig().SlotsPerEpoch.Mul(params.BeaconConfig().SecondsPerSlot))
+	deadline := genesisTime.Add(time.Duration(k.epochsToRun*secondsPerEpoch) * time.Second)
+
+	require.NoError(t, kw.RegisterPlaybooks(ctx), "Failed to register Assertoor playbooks")
+
+	if k.runSyncTest {
+		// Resume late-joining beacon node for normal sync and checkpoint sync test.
+		stoppedNodes, err := kw.StoppedPrysmCLName()
+		require.NoError(t, err, "Failed to locate the skip_start sync node")
+		require.Equal(t, 2, len(stoppedNodes))
+		require.Equal(t, true, slices.Contains(stoppedNodes, SYNC_NODE_SERVICE), "Expected stopped nodes to contain %s", SYNC_NODE_SERVICE)
+		require.Equal(t, true, slices.Contains(stoppedNodes, CHECKPOINT_SYNC_NODE_SERVICE), "Expected stopped nodes to contain %s", CHECKPOINT_SYNC_NODE_SERVICE)
+
+		delay := time.Until(genesisTime.Add(LATE_SYNC_NODE_DELAY))
+		scheduleLateSyncNodeStart(t, ctx, kw, delay, SYNC_NODE_SERVICE, CHECKPOINT_SYNC_NODE_SERVICE)
+	}
+
+	require.NoError(t, kw.WaitForAssertoor(ctx, deadline), "Assertoor checks failed")
+}
 
 // waitForNodeReady blocks until the beacon node reports healthy (200 from
 // /eth/v1/node/health) or ctx is done.
@@ -70,6 +155,37 @@ func fetchGenesisTime(t *testing.T, ctx context.Context, client *beacon.Client) 
 	require.NoError(t, err, "Failed to parse genesis time")
 
 	return time.Unix(secs, 0)
+}
+
+// scheduleLateSyncNodeStart starts the given skip_start beacon nodes after delay.
+func scheduleLateSyncNodeStart(t *testing.T, ctx context.Context, kw *kurtosis.KurtosisWrapper, delay time.Duration, names ...string) {
+	t.Logf("Will start late sync nodes %v after %s", names, delay)
+
+	done := make(chan error, len(names))
+	go func() {
+		select {
+		case <-ctx.Done():
+			return // run ended before the nodes were due to start
+		case <-time.After(delay):
+		}
+		for _, name := range names {
+			t.Logf("Starting late sync node %q", name)
+			done <- kw.StartService(name)
+		}
+	}()
+
+	t.Cleanup(func() {
+		// Non-blocking: report any start that actually ran and failed.
+		for range names {
+			select {
+			case err := <-done:
+				if err != nil {
+					t.Errorf("Failed to start late sync node: %v", err)
+				}
+			default:
+			}
+		}
+	})
 }
 
 // LoadPrysmDockerImages loads the Prysm beacon-chain and validator Docker images
